refactor: split the main app and site layout
