+++ apps/ui/src/components/Site/Layout.vue
@@ -1,0 +1,5 @@
+<template>
+  <SiteTopnav />
+  <router-view />
+  <SiteFooter />
+</template>
